+++ cmd/frontend/backend/symbols.go
@@ -3,6 +3,7 @@ package backend
 import (
 	"context"
 
+	"github.com/sourcegraph/sourcegraph/internal/search/search"
 	symbolsclient "github.com/sourcegraph/sourcegraph/internal/symbols"
 	"github.com/sourcegraph/sourcegraph/internal/symbols/protocol"
 )
@@ -13,7 +14,7 @@ var Symbols = &symbols{}
 type symbols struct{}
 
 // ListTags returns symbols in a repository from ctags.
-func (symbols) ListTags(ctx context.Context, args protocol.SearchArgs) ([]protocol.Symbol, error) {
+func (symbols) ListTags(ctx context.Context, args search.SymbolsParameters) ([]protocol.Symbol, error) {
 	result, err := symbolsclient.DefaultClient.Search(ctx, args)
 	if result == nil {
 		return nil, err