+++ cmd/frontend/internal/inventory/entries_test.go
@@ -12,7 +12,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/vcs/util"
 )
 
-func TestContext_Tree(t *testing.T) {
+func TestContext_Entries(t *testing.T) {
 	var (
 		readTreeCalls      []string
 		newFileReaderCalls []string
@@ -23,13 +23,13 @@ func TestContext_Tree(t *testing.T) {
 		ReadTree: func(ctx context.Context, path string) ([]os.FileInfo, error) {
 			readTreeCalls = append(readTreeCalls, path)
 			switch path {
-			case "":
+			case "d":
 				return []os.FileInfo{
-					&util.FileInfo{Name_: "a", Mode_: os.ModeDir},
-					&util.FileInfo{Name_: "b.go", Size_: 12},
+					&util.FileInfo{Name_: "d/a", Mode_: os.ModeDir},
+					&util.FileInfo{Name_: "d/b.go", Size_: 12},
 				}, nil
-			case "a":
-				return []os.FileInfo{&util.FileInfo{Name_: "a/c.m", Size_: 24}}, nil
+			case "d/a":
+				return []os.FileInfo{&util.FileInfo{Name_: "d/a/c.m", Size_: 24}}, nil
 			default:
 				panic("unhandled mock ReadTree " + path)
 			}
@@ -38,9 +38,11 @@ func TestContext_Tree(t *testing.T) {
 			newFileReaderCalls = append(newFileReaderCalls, path)
 			var data []byte
 			switch path {
-			case "b.go":
+			case "f.go":
+				data = []byte("package f")
+			case "d/b.go":
 				data = []byte("package main")
-			case "a/c.m":
+			case "d/a/c.m":
 				data = []byte("@interface X:NSObject {}")
 			default:
 				panic("unhandled mock ReadFile " + path)
@@ -59,43 +61,52 @@ func TestContext_Tree(t *testing.T) {
 		},
 	}
 
-	inv, err := c.Tree(context.Background(), &util.FileInfo{Name_: "", Mode_: os.ModeDir})
+	inv, err := c.Entries(context.Background(),
+		&util.FileInfo{Name_: "d", Mode_: os.ModeDir},
+		&util.FileInfo{Name_: "f.go", Mode_: 0, Size_: 1 /* HACK to force read */},
+	)
 	if err != nil {
 		t.Fatal(err)
 	}
 	if want := (Inventory{
 		Languages: []Lang{
+			{Name: "Go", TotalBytes: 13, TotalLines: 2},
 			{Name: "Objective-C", TotalBytes: 24, TotalLines: 1},
-			{Name: "Go", TotalBytes: 12, TotalLines: 1},
 		},
 	}); !reflect.DeepEqual(inv, want) {
 		t.Errorf("got  %#v\nwant %#v", inv, want)
 	}
 
 	// Check that our mocks were called as expected.
-	if want := []string{"", "a"}; !reflect.DeepEqual(readTreeCalls, want) {
+	if want := []string{"d", "d/a"}; !reflect.DeepEqual(readTreeCalls, want) {
 		t.Errorf("ReadTree calls: got %q, want %q", readTreeCalls, want)
 	}
 	if want := []string{
 		// We need to read all files to get line counts
-		"a/c.m",
-		"b.go",
+		"d/a/c.m",
+		"d/b.go",
+		"f.go",
 	}; !reflect.DeepEqual(newFileReaderCalls, want) {
 		t.Errorf("GetFileReader calls: got %q, want %q", newFileReaderCalls, want)
 	}
-	if want := []string{"", "a"}; !reflect.DeepEqual(cacheGetCalls, want) {
+	if want := []string{"d", "d/a", "f.go"}; !reflect.DeepEqual(cacheGetCalls, want) {
 		t.Errorf("CacheGet calls: got %q, want %q", cacheGetCalls, want)
 	}
 	if want := map[string]Inventory{
-		"": {
+		"d": {
 			Languages: []Lang{
-				{Name: "Objective-C", TotalBytes: 24, TotalLines: 1},
 				{Name: "Go", TotalBytes: 12, TotalLines: 1},
+				{Name: "Objective-C", TotalBytes: 24, TotalLines: 1},
 			},
 		},
-		"a": {
+		"d/a": {
 			Languages: []Lang{
 				{Name: "Objective-C", TotalBytes: 24, TotalLines: 1},
+			},
+		},
+		"f.go": {
+			Languages: []Lang{
+				{Name: "Go", TotalBytes: 1, TotalLines: 1},
 			},
 		},
 	}; !reflect.DeepEqual(cacheSetCalls, want) {