+++ web/src/e2e/e2e.test.ts
@@ -14,6 +14,7 @@ import { ExternalServiceKind } from '../../../shared/src/graphql/schema'
 import { getConfig } from '../../../shared/src/e2e/config'
 import * as assert from 'assert'
 import { asError } from '../../../shared/src/util/errors'
+import { Settings } from '../schema/settings.schema'
 
 const { gitHubToken, sourcegraphBaseUrl } = getConfig('gitHubToken', 'sourcegraphBaseUrl')
 
@@ -1367,6 +1368,50 @@ describe('e2e test suite', () => {
                 () => document.querySelectorAll('.e2e-literal-search-toast').length
             )
             expect(nodes).toEqual(0)
+        })
+    })
+
+    describe('Search statistics', () => {
+        beforeEach(async () => {
+            await driver.setUserSettings<Settings>({ experimentalFeatures: { searchStats: true } })
+        })
+        afterEach(async () => {
+            await driver.resetUserSettings()
+        })
+
+        test('button on search results page', async () => {
+            await driver.page.goto(`${sourcegraphBaseUrl}/search?q=abc`)
+            await driver.page.waitForSelector('a[href="/stats?q=abc"]')
+        })
+
+        test('page', async () => {
+            await driver.page.goto(`${sourcegraphBaseUrl}/stats?q=count%3A10000+abc`)
+
+            // Ensure the global navbar hides the search input (to avoid confusion with the one on
+            // the stats page).
+            await driver.page.waitForSelector('.global-navbar a.nav-link[href="/search"]')
+            assert.strictEqual(
+                await driver.page.evaluate(() => document.querySelectorAll('.e2e-query-input').length),
+                0
+            )
+
+            const queryInputValue = () =>
+                driver.page.evaluate(() => {
+                    const input = document.querySelector<HTMLInputElement>('.e2e-stats-query')
+                    return input ? input.value : null
+                })
+
+            // Check for a Go result (the sample repositories have Go files).
+            await driver.page.waitForSelector('a[href*="abc+lang:go"]')
+            assert.strictEqual(await queryInputValue(), 'count:10000 abc')
+            await percySnapshot(driver.page, 'Search stats')
+
+            // Update the query and rerun the computation.
+            await driver.page.type('.e2e-stats-query', 'd')
+            assert.strictEqual(await queryInputValue(), 'count:10000 abcd')
+            await driver.page.click('.e2e-stats-query-update')
+            await driver.page.waitForSelector('a[href*="abcd+lang:go"]')
+            assert.ok(driver.page.url().endsWith('/stats?q=count%3A10000+abcd'))
         })
     })