+++ web/src/search/input/QueryInput.tsx
@@ -18,7 +18,7 @@ import {
 import { eventLogger } from '../../tracking/eventLogger'
 import { scrollIntoView } from '../../util'
 import { Suggestion, SuggestionItem, createSuggestion, fuzzySearchFilters } from './Suggestion'
-import RegexpToggle from './RegexpToggle'
+import { RegexpToggle } from './RegexpToggle'
 import { PatternTypeProps } from '..'
 import Downshift from 'downshift'
 import { searchFilterSuggestions } from '../searchFilterSuggestions'